+++ CHANGELOG.md
@@ -18,10 +18,11 @@ This project adheres to
 
 ## [Unreleased](https://github.com/o1-labs/o1js/compare/045b1ab...HEAD)
 
-### Changed
+### Deprecated
 
-- Replace the `Circuit` API with the `ZkFunction` API, which provides the same
-  functionality with a more ergonomic and consistent structure, inspired by
+- Deprecate the `Circuit` API in favor of the `ZkFunction` API, now exposed via
+  the `Experimental` namespace (`Experimental.ZkFunction`), which provides the
+  same functionality with a more ergonomic and consistent structure, inspired by
   `ZkProgram`. https://github.com/o1-labs/o1js/pull/2245
 
 ### Fixed